Moved bids code into sct_compute_mtsat
Context: https://github.com/neuropoly/spinalcordtoolbox/pull/2875#discussion_r483385576
